requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.1
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
